requirements.txt
fastmcp==0.4.1
databricks-sql-connector==3.1.0
python-dotenv==1.0.1
